Created RichTextEditor component
