+++ backend/HexClanApplication.Api/bin/Debug/net5.0/HexClanApplication.Api.StaticWebAssets.xml
@@ -1,0 +1,1 @@
+﻿<StaticWebAssets Version="1.0" />
