+++ frontend/package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.26.1",
     "draft-js": "^0.11.7",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
